requirements.txt
PyExifTool
minio

